### 2025-11-29 8095386
shortened manuscript for joss
