+++ paper.bib
@@ -25,7 +25,6 @@
   langid = {english}
 }
 
-
 @book{decker2025,
   title = {The {{Computer Algebra System OSCAR}}: {{Algorithms}} and {{Examples}}},
   shorttitle = {The {{Computer Algebra System OSCAR}}},
