+++ paper.md
@@ -14,19 +14,15 @@ authors:
 affiliations:
  - name: Institute of Mathematics and Computer Science, University of Greifswald, Germany
    index: 1
-date: 27 November 2025
+date: 29 November 2025
 bibliography: paper.bib
 ---
 
 # Summary
-
 Singular perturbation theory is a mathematical toolbox that allows
 dimensionality reduction of ODE systems whose components evolve on different
 time scales emanating from the presence of a small parameter $\varepsilon$. 
-Intuitively, the fast components evolve so quickly that they can be approximated
-with their steady state on the slow time scale for $\varepsilon$ sufficiently
-small.
-More precisely, we get the limit
+More precisely, we obtain the *reduced system* as the limit
 \begin{equation}
   \label{eq:tikhonov}
   \begin{aligned}
@@ -39,186 +35,124 @@ More precisely, we get the limit
     0       & = h(u,v)
   \end{aligned}
 \end{equation}
-where the *reduced system* on the right is defined on the so-called *slow
-manifold* $M_0=\{(u,v)\,\vert\, h(u,v)=0\}$ and its dynamics can be described by
-the slow component $u$ alone (i.e. by local coordinates on $M_0$).
-This is known as Tikhonov's [-@tikhonov1952] theorem [see @verhulst2007 for its
-present-day form]. 
-Dimensionality reduction is very useful for model derivation and analysis, but
-requiring the system to be in standard form as in \autoref{eq:tikhonov} can make
-it difficult to find meaningful reductions, as it relies on the choice of
-appropriate coordinates. 
-The importance of a coordinate-free approach was already pointed out by
-@fenichel1979, who established a geometric singular perturbation theory (GSPT;
-see @wechselberger2020 for an overview).
-More recently, Goeke and Walcher (together with colleagues) developed an
-algebraic approach to GSPT that allows to find all critical parameters admitting
-a reduction in the sense of Tikhonov and Fenichel together with their slow
-manifolds (implicitly given as affine varieties) systematically for polynomial
-or rational ODE systems [@goeke2013; @goeke2013a; @goeke2014; @goeke2015]. 
-These critical parameters define a slow-fast separation of processes instead of
-components, i.e. we obtain reductions for systems of the form
+as stated in Tikhonov's theorem [-@tikhonov1952; Theorem 1.1 in @verhulst2007].
+For autonomous systems, @fenichel1979 established a geometric singular
+perturbation theory (GSPT) in coordinate-free settings [see @wechselberger2020].
+Convergence properties then follow from the slow manifold 
+$M_0=\{(u,v)\,\vert\,h(u,v)=0\}$, on which the reduced system is defined.
+
+The algebraic approach to GSPT recently developed by Goeke and Walcher (and
+colleagues) allows to systematically find all critical parameters admitting a
+reduction for polynomial or rational ODE systems of the form
 \begin{equation}
     \label{eq:slowfastsystem}
     \dot{x} = f(x,\pi) = f^{(0)}(x) + \varepsilon f^{(1)}(x) + \mathcal{O}(\varepsilon^2), 
     \quad x\in\mathbb{R}^n, \pi\in\mathbb{R}^m,
 \end{equation}
+i.e. we obtain all reductions for a system with a slow-fast separation of
+processes instead of components as in the standard form (\ref{eq:tikhonov}),
 which renders this a coordinate-free approach.
-The main idea is to evaluate necessary conditions for a slow-fast separation as
-in \autoref{eq:slowfastsystem} for the existence of a reduction.
+This can be achieved by evaluating necessary conditions for the existence of a
[…]
