### 2025-12-02 7cc8972
updated installation: TFR is in general registry now
