### 2025-12-14 #9527 @nkrusch
:wave: Hello @ranocha, @DanielVandH, @avik-pal, @chaypaterson, and @highlando,

would you be interested in reviewing this submission for JOSS?
