### 2026-01-06 #9527 @nkrusch
:wave: Hello @gchure, @mbudisic, @kaipaine, and @dawbarton

would you be interested in reviewing this submission for JOSS?
