### 2026-01-07 a2df3b5
fixed bug with compound fractions in automatic computation of explicit manifold
