### 2026-01-07 #9527 @kaipaine
I'm not familiar enough with the theory to review this submission.
