### 2026-01-14 dbe3437
updated manuscript for JOSS
