+++ paper.md
@@ -14,7 +14,7 @@ authors:
 affiliations:
  - name: Institute of Mathematics and Computer Science, University of Greifswald, Germany
    index: 1
-date: 29 November 2025
+date: 14 January 2026
 bibliography: paper.bib
 ---
 
@@ -65,7 +65,7 @@ The ad-hoc approach to singular perturbation theory requires prior knowledge
 about a suitable time scale separation and substantial mathematical effort to
 compute the reduction.
 The algebraic approach yields algorithmically accessible conditions for the
-existence of a reduction, which allows us to find all reductions of a given
+existence of a reduction, which allows to find *all* reductions of a given
 polynomial ODE system using methods from computational algebra, and simplifies
 the computation of reduced systems [@goeke2015].
 `TikhonovFenichelReductions.jl` makes the required computations easily
@@ -76,63 +76,63 @@ The author is not aware of any publicly available implementation of the theory
 by Goeke and Walcher, but there exist an implementation for a related approach
 for computing invariant manifolds by Roberts [-@roberts1997; -@roberts-webapp].
 
+# Software design 
+`TikhonovFenichelReductions.jl` is implemented in `Julia` due to its
+flexibility, the use of multiple dispatch and the availability of the
+feature-rich computer algebra system `Oscar.jl` [@oscar2022; @decker2025]. 
+
+Core features are the search for critical parameters admitting a reduction,
+so-called *Tikhonov-Fenichel Parameter Values (TFPVs)*, and the computation of
+the corresponding reduced systems.
+Crucially, this requires various computations with multiple symbolic
+representations (i.e. different polynomial rings, rational function fields and
+matrix spaces) and therefore parsing of data between the corresponding types in
+`Oscar.jl`, which `TikhonovFenichelReductions.jl` performs hidden away from the
+user. 
+Thus, the user mostly works in an object-oriented manner with the types and
+methods provided.
+
+The package is essentially designed around two types:
+`ReductionProblem`, which constructs all symbolic data types needed for the
+search of TFPVs, 
+and `Reduction`, which holds all relevant information for the reduced
+system and the steps required to compute it.
+The latter also contains the reduced system and other information parsed to the
+appropriate types from `Oscar.jl`, that can be further used, e.g. for a symbolic
+analysis.
+
 # Features
-The main features provided by `TikhonovFenichelReductions.jl` are the search for
-critical parameters admitting a reduction, so-called *Tikhonov-Fenichel
-Parameter Values (TFPVs)*, and the computation of the corresponding reduced
-systems.
 Detailed explanations and examples are provided by @apelt2025 and in the
 [documentation](https://jo-ap.github.io/TikhonovFenichelReductions.jl/stable/).
 
 ## Finding TFPVs
-The package provides two methods for finding TFPVs admitting a reduction onto an
-$s$-dimensional slow manifold.
-
-Restricting the search for TFPVs to *slow-fast separations of rates*, i.e. TFPVs
-with some parameters set to zero, is typically more efficient and yields the
-TFPV candidates together with their slow manifolds given implicitly as the
-irreducible components of the corresponding affine variety
-$\mathcal{V}(f^{(0)})=\{x\in\mathbb{R}^n \mid f^{(0)}(x)=0\}$.
-
-Computing a Gröbner basis $G$ with an elimination ordering for the state
-variables of an ideal obtained from necessary conditions for the existence of a
-reduction yields all TFPVs implicitly as $\mathcal{V}(G)$.
-This may be infeasible to compute, but can reveal more complex expressions that
-can be considered as small parameters.
-However, introducing new parameters from these expressions typically allows us
[…]
