+++ paper.bib
@@ -1,3 +1,17 @@
+@article{abbott2020,
+  title = {Fast and Slow Advances toward a Deeper Integration of Theory and Empiricism},
+  author = {Abbott, Karen C. and Ji, Fang and Stieha, Christopher R. and Moore, Christopher M.},
+  date = {2020-03},
+  journaltitle = {Theoretical Ecology},
+  shortjournal = {Theor Ecol},
+  volume = {13},
+  number = {1},
+  pages = {7--15},
+  issn = {1874-1738, 1874-1746},
+  doi = {10.1007/s12080-019-00441-x},
+  langid = {english}
+}
+
 @article{apelt2025,
   title = {Tikhonov-{{Fenichel}} Reductions and Their Application to a Novel Modelling Approach for Mutualism},
   author = {Apelt, Johannes and Liebscher, Volkmar},
@@ -106,6 +120,20 @@
   langid = {english}
 }
 
+@article{kruff2019,
+  title = {The {{Rosenzweig}}–{{MacArthur}} System via Reduction of an Individual Based Model},
+  author = {Kruff, Niclas and Lax, Christian and Liebscher, Volkmar and Walcher, Sebastian},
+  date = {2019-01},
+  journaltitle = {Journal of Mathematical Biology},
+  shortjournal = {J. Math. Biol.},
+  volume = {78},
+  number = {1--2},
+  pages = {413--439},
+  issn = {0303-6812, 1432-1416},
+  doi = {10.1007/s00285-018-1278-y},
+  langid = {english}
+}
+
 @article{loman2023,
   title = {Catalyst: {{Fast}} and Flexible Modeling of Reaction Networks},
   shorttitle = {Catalyst},
@@ -120,6 +148,33 @@
   publisher = {Public Library of Science (PLoS)},
   issn = {1553-7358},
   doi = {10.1371/journal.pcbi.1011530},
+  langid = {english}
+}
+
+@article{poggiale2004,
+  title = {Impact of Spatial Heterogeneity on a Predator–Prey System Dynamics},
+  author = {Poggiale, Jean-Christophe and Auger, Pierre},
+  date = {2004-11-01},
+  journaltitle = {Comptes Rendus. Biologies},
+  volume = {327},
+  number = {11},
+  pages = {1058--1063},
+  issn = {1768-3238},
+  doi = {10.1016/j.crvi.2004.06.006},
+  langid = {english}
+}
+
+@article{revilla2015,
+  title = {Numerical Responses in Resource-Based Mutualisms: {{A}} Time Scale Approach},
+  shorttitle = {Numerical Responses in Resource-Based Mutualisms},
+  author = {Revilla, Tomás A.},
+  date = {2015-08},
+  journaltitle = {Journal of Theoretical Biology},
+  shortjournal = {Journal of Theoretical Biology},
+  volume = {378},
+  pages = {39--46},
+  issn = {00225193},
+  doi = {10.1016/j.jtbi.2015.04.012},
   langid = {english}
 }
 
