### 2026-01-14 #9527 @nkrusch
@danielskatz this submission is in good shape, with open development standards, clearly in scope by prior criteria, and I have not spotted any concerns relating to AI. As I do not have concerns, I would like to proceed with the reviewer search.
