### 2026-01-14 #9527 @gchure
Hi @nkrusch, sorry for the late response. This paper looks interesting but I don't have the bandwidth to review right now.
