### 2026-01-17 #9527 @nkrusch
Hello @abhijit-c, @mongibellili, @mauro3, @epspebble and @ysimillides

would you be interested in reviewing this submission for JOSS?
