### 2026-01-18 #9527 @oskarhenriksson
@nkrusch It looks like an interesting paper that touches on many of my mathematical interests, so I would be happy to review it. However, I won't have time do it until around mid-February.
