### 2026-01-19 #9527 @mongibellili
Greetings,
The subject is outside my area of expertise, and I am currently busy. I can
review the computer science side starting April if you want.

On Mon, Jan 19, 2026, 8:30 AM The Open Journals editorial robot <
***@***.***> wrote:

> *editorialbot* left a comment (openjournals/joss-reviews#9527)
> <https://github.com/openjournals/joss-reviews/issues/9527#issuecomment-3766866967>
>
> @oskarhenriksson <https://github.com/oskarhenriksson> added to the
> reviewers list!
>
