### 2026-01-22 97c3bfe
added condition for early stopping in get_explicit_manifold
