+++ Project.toml
@@ -1,7 +1,7 @@
 name = "TikhonovFenichelReductions"
 uuid = "66629163-d46e-44ca-9a00-7025b1e168e6"
 authors = ["jo-ap"]
-version = "0.3.3"
+version = "0.3.4"
 
 [deps]
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
