### 2026-01-23 f77089e
use julia workspaces for dependencies
