+++ test/Project.toml
@@ -3,3 +3,4 @@ Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 Catalyst = "479239e8-5488-4da2-87a7-35f2df7eef83"
 Coverage = "a2441757-f6aa-5fb2-8edb-039e3f45d037"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+TikhonovFenichelReductions = "66629163-d46e-44ca-9a00-7025b1e168e6"
