+++ Project.toml
@@ -26,3 +26,6 @@ Logging = "1.9"
 Oscar = "1.0.0 - 1.2.1, 1.3.1, 1.4"
 Pkg = "1.9"
 julia = "1.9"
+
+[workspace]
+projects = ["docs", "test"]
