+++ v0.3.4/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"gettingstarted/#Getting-Started","page":"Getting Started","title":"Getting Started","text":"","category":"section"},{"location":"gettingstarted/#Installation","page":"Getting Started","title":"Installation","text":"Run\n\nadd TikhonovFenichelReductions\n\nin Julia package Mode.\n\nnote: Installation\nTikhonovFenichelReductions.jl relies on Oscar.jl, which requieres at least 6GB of free memory for the installation (the Oscar-Team recommends at least 16GB). On Windows, Oscar.jl needs to be installed using the Windows Subsystem for Linux (WSL). Instructions can be found in their documentation.","category":"section"},{"location":"gettingstarted/#Example","page":"Getting Started","title":"Example","text":"Here we consider the derivation of the Rosenzweig-MacArthur model via a reduction from a three dimensional system as demonstrated in [8]. \n\nLoad the package and its dependency Oscar.jl.  Note that loading Oscar is optional, but results in pretty printing of types and imports useful functions.\n\nusing TikhonovFenichelReductions\nusing Oscar # optional\n\nDefine the components, parameters and the RHS of the system. \n\n# components\nx = [\"B\", \"S\", \"H\"]\n\n# parameters\np = [\"α\", \"β\", \"γ\", \"δ\", \"η\", \"ρ\"]\n\n# RHS of ODE system ẋ = f(x, p), where f is polynomial in x and p\nfunction f(x, p)\n  B, S, H = x\n  α, β, γ, δ, η, ρ = p\n  return [\n    ρ*B*(1-B) - α*B*H,\n    -η*S + γ*B*H,\n    β*S- δ*H + η*S - γ*B*H\n  ]\nend\n\nInitialize the problem and set the desired dimension of the reduced system\n\n# dimension of the reduced system\ns = 2\n\n# create problem\nproblem = ReductionProblem(f, x, p, s)","category":"section"},{"location":"gettingstarted/#Slow-Fast-Separations-of-Rates","page":"Getting Started","title":"Slow-Fast Separations of Rates","text":"We can find all slow-fast separations that are TFPVs by using necessary conditions for the existence of a reduction.  This returns a vector of boolean indices for each TFPV candidate, where 0 corresponds to a small parameters, i.e. those multiplied by varepsilon, and 1 corresponds to a parameter in mathcalO(1). In addition, we obtain the generators of the irreducible components of the affine variety \n\nmathcalV_mathbbC(f(cdotpi^star)) = xinmathbbC^n mid f(xpi^star)=0\n\nand their dimension, where pi^star is defined by the corresponding slow-fast separation in tfpvs.  This information is stored using the type Variety. Note that we later have to check manually whether the variety taken in mathbbR^n has the same dimension (i.e. if there exists a real non-singular point), which then at least locally renders this variety a manifold.\n\ntfpvs, varieties = tfpvs_and_varieties(problem);\n\nShow the results: All possible slow-fast separation of rates \n\nprint_tfpvs(problem, tfpvs)\n\nand the corresponding irreducible components containing the slow manifold and their Krull dimension\n\nprint_varieties(varieties)\n\nFor further use, we need to make the variables (the system components and parameters) available in the Main namespace. Then, we can compute the reduction corresponding to TFPV 15, which is the Rosenzweig-MacArthur model.\n\nB, S, H = system_components(problem)\nα, β, γ, δ, η, ρ = system_parameters(problem)\n\n# instantiate reduction 
[…]
