+++ v0.3.4/refs.bib
@@ -1,0 +1,89 @@
+@article{apelt2025,
+  title = {Tikhonov-{{Fenichel}} Reductions and Their Application to a Novel Modelling Approach for Mutualism},
+  author = {Apelt, Johannes and Liebscher, Volkmar},
+  year = {2025},
+  journal = {Theoretical Population Biology},
+  pages = {16--35},
+  doi = {10.1016/j.tpb.2025.08.004},
+  langid = {english}
+}
+
+@article{fenichel1979,
+  title = {Geometric Singular Perturbation Theory for Ordinary Differential Equations},
+  author = {Fenichel, Neil},
+  year = {1979},
+  journal = {Journal of Differential Equations},
+  volume = {31},
+  number = {1},
+  pages = {53--98},
+  doi = {10.1016/0022-0396(79)90152-9},
+  }
+
+@incollection{goeke2013a,
+  title = {Quasi-{{Steady State}}: {{Searching}} for and {{Utilizing Small Parameters}}},
+  shorttitle = {Quasi-{{Steady State}}},
+  booktitle = {Recent {{Trends}} in {{Dynamical Systems}}},
+  author = {Goeke, Alexandra and Walcher, Sebastian},
+  editor = {Johann, Andreas and Kruse, Hans-Peter and Rupp, Florian and Schmitz, Stephan},
+  year = {2013},
+  volume = {35},
+  pages = {153--178},
+  publisher = {{Springer Basel}},
+  location = {{Basel}},
+  doi = {10.1007/978-3-0348-0451-6_8},
+  isbn = {978-3-0348-0450-9},
+}
+
+@article{goeke2014,
+  title = {A Constructive Approach to Quasi-Steady State Reductions},
+  author = {Goeke, Alexandra and Walcher, Sebastian},
+  year = {2014},
+  journal = {Journal of Mathematical Chemistry},
+  volume = {52},
+  number = {10},
+  pages = {2596--2626},
+  doi = {10.1007/s10910-014-0402-5},
+  }
+
+@article{goeke2015,
+  title = {Determining "Small Parameters" for Quasi-Steady State},
+  author = {Goeke, Alexandra and Walcher, Sebastian and Zerz, Eva},
+  year = {2015},
+  journal = {Journal of Differential Equations},
+  volume = {259},
+  number = {3},
+  pages = {1149--1180},
+  doi = {10.1016/j.jde.2015.02.038},
+}
+
+@article{tikhonov1952,
+  title = {Systems of differential equations containing small parameters in the derivatives},
+  author = {Tikhonov, Andrei Nikolaevich},
+  year = {1952},
+  journal = {Matematicheskii sbornik},
+  volume = {73},
+  number = {3},
+  pages = {575--586},
+  }
+
+@article{kruff2019,
+  title = {The {{Rosenzweig}}--{{MacArthur}} System via Reduction of an Individual Based Model},
+  author = {Kruff, Niclas and Lax, Christian and Liebscher, Volkmar and Walcher, Sebastian},
+  year = {2019},
+  journal = {Journal of Mathematical Biology},
+  volume = {78},
+  number = {1-2},
+  pages = {413--439},
+  doi = {10.1007/s00285-018-1278-y},
+  }
+
+@article{verhulst2007,
+  title = {Singular Perturbation Methods for Slow--Fast Dynamics},
+  author = {Verhulst, Ferdinand},
+  year = {2007},
+  journal = {Nonlinear Dynamics},
+  volume = {50},
+  number = {4},
+  pages = {747--753},
+  doi = {10.1007/s11071-007-9236-z},
+}
