### 2026-01-26 #9527 @nkrusch
👋 @jiweiqi, @YingboMa, and @njchiang,

would you be interested in reviewing this submission for JOSS?
