### 2026-01-29 #9527 @nkrusch
:wave: @jredondoyuste would you be interested in reviewing this submission for JOSS?
