### 2026-01-30 #9527 @nkrusch
:wave: @HectorMozo3110 and @sandreza,

would you be interested in reviewing this submission for JOSS?
