### 2026-02-02 #9924 @editorialbot
[REVIEW]: TikhonovFenichelReductions.jl: A systematic approach to geometric singular perturbation theory
**Submitting author:** @jo-ap (<a href="http://orcid.org/0009-0000-8973-7495">Johannes Apelt</a>)
**Repository:** https://github.com/jo-ap/TikhonovFenichelReductions.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.5
