### 2026-02-02 #9924 @nkrusch
Hi @jo-ap,

Currently the license info check is returning:

> 🟡 License found: Other ([Check here](https://opensource.org/licenses) for OSI approval)

GitHub is also having some difficulty correctly recognizing the license.
See e.g., "View license" in the about section, as opposed to saying "GPL-3.0 license".

The reason is that the [license file has some extra text at the beginning of the file](https://github.com/jo-ap/TikhonovFenichelReductions.jl/blob/f77089e1a798dcc4cce76d756270ae9dfe1af98d/LICENSE#L1-L12).

The license file should begin with "GNU GENERAL PUBLIC LICENSE...", just as in:
https://www.gnu.org/licenses/gpl-3.0.txt

Lastly, here is a nice guide to using GNU Licenses in general https://www.gnu.org/licenses/gpl-howto.html
