### 2026-02-02 f8cd215
rearranged terms in Rosenzweig-MacArthur model
