### 2026-02-02 #9924 @jo-ap
Hi @nkrusch, thanks, I removed the extra text from the license. GitHub now seems to recognize it correctly.
