### 2026-02-02 #9924 @nkrusch
The result with the editorialbot is now as good as possible.

The license detection is based on a package, and appears there are [some challenges](https://github.com/licensee/licensee/issues/631) with handling the different variants of the GNU family, which explains the warning.

I am satisfied that the license has been applied correctly.
