### 2026-02-28 jo-ap/TikhonovFenichelReductions.jl#1 @HectorMozo3110
Suggested README Improvements (JOSS Review)
Hi 👋 @jo-ap,

I’ve been taking a look at the README and I noticed a few things that could potentially be added or clarified to improve accessibility and overall structure.

First, the Overview section could be slightly expanded. It already explains what the package does, but it might help to add a bit more high-level context about the type of problems it addresses, such as polynomial ODE systems, algebraic reduction, and timescale separation. It could also briefly mention the intended audience, for example researchers working in dynamical systems, mathematical biology, chemical reaction networks, or symbolic computation. Just a few additional lines would make it clearer for readers who may not already be familiar with the methodology.

It might also be helpful to add a clearly labeled Requirements section. Even though some requirements are mentioned, listing them explicitly, such as the minimum Julia version, the dependency on Oscar.jl, and the WSL requirement for Windows users, would make installation expectations more transparent.

The Installation section could also be slightly expanded to show the full workflow and possibly include a short note about symbolic dependencies, so users know what to expect during setup.

I noticed there is a documentation badge, which is great. However, it might be even better to include a dedicated Documentation section in the README with direct clickable links to the stable and development documentation. That way users can immediately access the docs from a clearly structured section without relying only on badges.

If tests are already configured, it could also be useful to add a short Running Tests section with the standard command and possibly a link to the test directory.

These are my observations specifically regarding the README. I will provide additional comments about other parts of the repository or the paper in a separate thread. For now, these are my comments concerning the README. I remain attentive to your response.

**Note:** This observation is part of the ongoing JOSS review (see https://github.com/openjournals/jossreviews/issues/https://github.com/openjournals/joss-reviews/issues/9924)
