### 2026-03-03 e6718fb
improved README (fixes #1)
