### 2026-03-03 #9527 @nkrusch
@oskarhenriksson - no problem. Once you are free, you will find the review thread awaiting here:
https://github.com/openjournals/joss-reviews/issues/9924
