+++ README.md
@@ -2,8 +2,7 @@
 
 [![](https://img.shields.io/badge/docs-stable-blue.svg)](https://jo-ap.github.io/TikhonovFenichelReductions.jl/stable)
 [![](https://img.shields.io/badge/docs-dev-blue.svg)](https://jo-ap.github.io/TikhonovFenichelReductions.jl/dev)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14411664.svg)](https://doi.org/10.5281/zenodo.14411664)
-
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.18352148.svg)](https://doi.org/10.5281/zenodo.18352148)
 
 A Julia package for computing Tikhonov-Fenichel Parameter Values (TFPVs) for
 polynomial ODE systems and their corresponding reductions (see [1-3] for
@@ -62,7 +61,7 @@ More details can be found
 [here](https://jo-ap.github.io/TikhonovFenichelReductions.jl/stable/gettingstarted/#Installation) 
 and in the documentation for [Oscar.jl](https://www.oscar-system.org/).
 
-## Development and Tests
+## Contribution
 Please open an issue here, if you encounter a bug or want to share suggestions
 for the package.
 Pull requests are also welcome, preferably after reaching out. 
