### 2026-03-04 fb17abb
Update DOI badge to point to last release
Updated DOI badge in README.md to a new identifier.
