### 2026-03-21 #9924 @nkrusch
@oskarhenriksson this is the review thread of the submissions. Your first step is to generate a personalized checklist. To do so, enter `@editorialbot generate my checklist` as a comment in this thread.
