### 2026-04-05 jo-ap/TikhonovFenichelReductions.jl#2 @oskarhenriksson
Questions about tests (JOSS review)
Hi @jo-ap,

I was looking over the test files and have a couple of small questions and suggestions that I wanted to get your thoughts on as part of the review for JOSS (see https://github.com/openjournals/joss-reviews/issues/9924):

1) As far as I can tell, the tests are essentially all focused on a single example, namely the classical Michaelis–Menten system. Would it be useful to also include some tests based on a **larger and more complex example**, such as the polynomial system in `example.jl` or the rational system you studied in https://doi.org/10.1016/j.tpb.2025.08.004? Alternatively (or additionally), some of the examples shown in the documentation could perhaps be turned into doctests?

2) Are there any **edge cases** that could be relevant to include in the test suite? (Just as an example: does something interesting that needs to be tested happen if the user inputs a system of differential equations without positive steady states, or systems where the Jacobian is singular at all steady states?)

3) Are there any bugs that were fixed during the development process that should be covered by **regression tests**?

4) Have you considered running your test suite automatically on pull requests and/or commits to main via CI and GitHub Actions as a protection of the main branch?
