### 2026-04-05 jo-ap/TikhonovFenichelReductions.jl#3 @oskarhenriksson
Tiny bug in `example.jl` (JOSS review)
When I run [example.jl](https://github.com/jo-ap/TikhonovFenichelReductions.jl/blob/main/example.jl), I get an error at line 75:

```julia-repl
julia> reduction = Reduction(problem, sf_separations[15])
ERROR: BoundsError: attempt to access 6-element Vector{Vector{Bool}} at index [15]
Stacktrace:
 [1] throw_boundserror(A::Vector{Vector{Bool}}, I::Tuple{Int64})
   @ Base ./essentials.jl:15
 [2] getindex(A::Vector{Vector{Bool}}, i::Int64)
   @ Base ./essentials.jl:919
 [3] top-level scope
   @ REPL[13]:1
```

The reduction you intended to showcase seems to be the one with index 6, not 15.

I'm testing this as part of the ongoing JOSS review https://github.com/openjournals/joss-reviews/issues/9924.
