### 2026-04-05 jo-ap/TikhonovFenichelReductions.jl#4 @oskarhenriksson
Documentation of mutation functions (JOSS review)
The mutation functions `compute_reduction!`, `set_decomposition!` and `set_manifold!` can fail if given inappropriate input, and therefore include some verifying checks. The result of these checks is indicated by a Boolean that is returned by the functions. The interpretation of this is already clarified by `@info` statements, but could it perhaps also make sense to explain in a bit more detail in the documentation what these checks are and what exactly `true`/`false` mean?

This is part of https://github.com/openjournals/joss-reviews/issues/9924.
