### 2026-04-07 4011fc3
compute all reductions for example (fixes #3)
