### 2026-04-07 jo-ap/TikhonovFenichelReductions.jl#5 @oskarhenriksson
Target audience (JOSS review)
The review guidelines for the JOSS review https://github.com/openjournals/joss-reviews/issues/9924 ask me to check whether the documentation and the "Statement of Need" section of the paper "clearly state who the target audience is". I don't see any explicit mention of this in either the docs or the Statement of Need, so I wanted to ask if you could consider adding it.
