### 2026-04-07 jo-ap/TikhonovFenichelReductions.jl#6 @oskarhenriksson
Where to turn for support? (JOSS review)
The review guidelines for https://github.com/openjournals/joss-reviews/issues/9924 ask me to check whether there are clear guidelines for "third parties wishing to [...] seek support". The contribution guidelines in the README tell the user to raise an issue if they encounter a bug or have suggestions – but what should they do if they have a question about how the package works? Should they raise an issue? Alternatively (or additionally), one could enable the "Discussions" tab in the repository and direct users there for questions. Regardless of which policy you want to implement, I think it would be helpful to say something explicit about where to direct questions in the README.
