### 2026-04-08 433ed55
defined equality for `ReductionProblem` and `Reduction`
