### 2026-04-08 49c9885
fixed f0 being parsed incorrectly for constant RHS (fixes #9)
