+++ Project.toml
@@ -18,12 +18,12 @@ Catalyst = "479239e8-5488-4da2-87a7-35f2df7eef83"
 CatalystExt = "Catalyst"
 
 [compat]
-Catalyst = ">=14"
+Catalyst = "16"
 DocStringExtensions = "0.9"
 LaTeXStrings = "1.3"
 Latexify = "0.16"
 Logging = "1.9"
-Oscar = "1.0.0 - 1.2.1, 1.3.1, 1.4"
+Oscar = "1.4"
 Pkg = "1.9"
 julia = "1.9"
 
