### 2026-04-08 6d1e577
fixed printing for reduced system
