### 2026-04-08 e2dfaed
added more tests (fixes #2), including example from #9
