### 2026-04-08 df6adb9
added automated testing to CI
