### 2026-04-09 fdb8fa0
make _get_varieties return irreducible components as Vector{Variety} instead of generating ideals
