### 2026-04-09 f1b6d3d
added real world examples (fixes #8)
