+++ docs/make.jl
@@ -11,7 +11,7 @@ bib = CitationBibliography(
 makedocs(
   sitename="TikhonovFenichelReductions.jl",
   repo=Remotes.GitHub("jo-ap", "TikhonovFenichelReductions.jl"),
-  pages=["index.md", "gettingstarted.md", "hints.md", "api.md", "references.md"],
+  pages=["index.md", "gettingstarted.md", "examples.md", "hints.md", "api.md", "references.md"],
   clean=true;
   plugins=[bib]
 )
