+++ docs/src/examples.md
@@ -1,0 +1,198 @@
+# Examples
+
+## Michaelis-Menten kinetics 
+The Michaelis-Menten reaction scheme
+```math
+E + S \xrightleftharpoons[k_{-1}]{k_1} C \xrightleftharpoons[k_{-2}]{k_2} E + P
+```
+leads to the following ODE system (using the conserved quantities $E + C$ and
+$S + C + P$)
+```math
+\begin{aligned}
+    \frac{dS}{dt} &= -k_1 e_0 S + (k_1 S + k_{-1}) C \\ 
+    \frac{dC}{dt} &= k_1 e_0 S - (k_1 S + k_{-1} + k_2) C + k_{-2} (e_0 - C) (s_0 - S - C)
+\end{aligned}
+```
+where $e_0 = E(0), s_0 = S(0)$.
+
+The classical one-dimensional equation for the Michaelis-Menten reaction
+kinetics is typically derived using a quasi-steady state (QSS) assumption.
+Following [goeke2013a](@cite), this approach can be formalized with
+Tikhonov-Fenichel reduction theory.
+Here, we see how this can be achieved using `TikhonovFenichelReductions.jl`.
+In the following, we consider the irreversible and reversible case, i.e.
+$k_{-2}=0$ and $k_{-2}>0$, respectively.
+
+```@example 1
+# Setup
+using Oscar 
+using TikhonovFenichelReductions
+
+# Define the system
+x = ["S", "C"]
+p = ["e₀", "s₀", "k₁", "k₋₁", "k₂", "k₋₂"]
+function f(x, p)
+  S, C = x
+  e₀, s₀, k₁, k₋₁, k₂, k₋₂ = p
+  return [-k₁*e₀*S + (k₁*S + k₋₁)*C, k₁*e₀*S - (k₁*S + k₋₁ + k₂)*C + k₋₂*(e₀ - C)*(s₀ - S - C)]
+end
+``` 
+### The irreversible case
+We start by considering the irreversible case, i.e. $k_{-2} = 0$.
+```@example 1
+# RHS for the irreversible case
+f1(x,p) = f(x, p .* [1,1,1,1,1,0])
+
+# Initialize the problem
+problem = ReductionProblem(f1, x, p, 1)
+```
+```@example 1
+# Search for TFPVs (fix s₀ and k₋₂, since they are irrelevant)
+tfpvs, varieties = tfpvs_and_varieties(problem; preset=(s₀ = 1, k₋₂ = 1));
+print_tfpvs(problem, tfpvs)
+```
+
+```@example 1 
+print_varieties(varieties)
+```
+
+We can find all unique manifolds and compute the corresponding reductions.
+```@example 1 
+unique_V = unique_varieties(problem, varieties)
+M = [get_explicit_manifold(problem, V) for V in unique_V]
+```
+
+```@example 1 
+@assert all([m[2] for m in M])
+R, idx = compute_reductions(problem, tfpvs, varieties, unique_V, [m[1] for m in M])
+```
+
+The ad-hoc method to reduce this system found in most textbooks relies on the
+assumption that (at least in the initial phase of the reaction)
+$\frac{dC}{dt}\in\mathcal{O}(\varepsilon)$ for $\varepsilon>0$ small.
+Thus, one approximates $\varepsilon=0$ and solves
+```math 
+0 = \frac{dC}{dt} = k_1 e_0 S - (k_1 S + k_{-1} + k_2) C,
+``` 
+for $C$, which yields 
+```math 
+C = \frac{k_1 e_0 S}{k_1 S + k_{-1} + k_2}
+```
+and thus 
+```math 
[…]
