### 2026-04-09 a4def3a
linked to discussions for support and questions (fixes #6)
