### 2026-04-10 a3e159b
improved description of checks for mutating functions (fixes #4)
