### 2026-04-10 jo-ap/TikhonovFenichelReductions.jl#4 @jo-ap
Hi @oskarhenriksson,
I have expanded the explanations for the mutating functions you mentioned on the getting started page of the docs. During this, I also spotted that I have not clarified that bulk computation of reductions relies on `set_decomposition!` to work with only passing the variety. If that fails, a warning is thrown and a statement is shown in the docstring, but it is now also prominently shown on the [getting started page](https://jo-ap.github.io/TikhonovFenichelReductions.jl/dev/gettingstarted/).
Again, please reopen the issue if you think this is to brief.
