### 2026-04-28 #9924 @danielskatz
👋 @nkrusch  - As track chair, I try to check on submissions in the CSISM track semi-regularly. For this one, what's its status? I can't see that anything has happened in the last few weeks.
