### 2026-04-28 #9924 @nkrusch
@danielskatz thanks for checking in! I believe we are waiting for a paper update from the author (cf. issue [#7](https://github.com/jo-ap/TikhonovFenichelReductions.jl/issues/7)).

@jo-ap, please give an estimate of when you expect the update to be ready.
