### 2026-05-04 845b165
updated manuscript for JOSS
