+++ paper.bib
@@ -120,6 +120,20 @@
   langid = {english}
 }
 
+@article{hek2010,
+  title = {Geometric Singular Perturbation Theory in Biological Practice},
+  author = {Hek, Geertje},
+  date = {2010-03},
+  journaltitle = {Journal of Mathematical Biology},
+  shortjournal = {J. Math. Biol.},
+  volume = {60},
+  number = {3},
+  pages = {347--386},
+  issn = {0303-6812, 1432-1416},
+  doi = {10.1007/s00285-009-0266-7},
+  langid = {english}
+}
+
 @article{kruff2019,
   title = {The {{Rosenzweig}}–{{MacArthur}} System via Reduction of an Individual Based Model},
   author = {Kruff, Niclas and Lax, Christian and Liebscher, Volkmar and Walcher, Sebastian},
