### 2026-05-07 bbc5168
Manuscript edits (JOSS review)
