+++ paper/paper.bib
@@ -1,0 +1,260 @@
+@article{abbott2020,
+  title = {Fast and Slow Advances toward a Deeper Integration of Theory and Empiricism},
+  author = {Abbott, Karen C. and Ji, Fang and Stieha, Christopher R. and Moore, Christopher M.},
+  date = {2020-03},
+  journaltitle = {Theoretical Ecology},
+  shortjournal = {Theor Ecol},
+  volume = {13},
+  number = {1},
+  pages = {7--15},
+  issn = {1874-1738, 1874-1746},
+  doi = {10.1007/s12080-019-00441-x},
+  langid = {english}
+}
+
+@article{apelt2025,
+  title = {Tikhonov-{{Fenichel}} Reductions and Their Application to a Novel Modelling Approach for Mutualism},
+  author = {Apelt, Johannes and Liebscher, Volkmar},
+  date = {2025-09},
+  journaltitle = {Theoretical Population Biology},
+  shortjournal = {Theoretical Population Biology},
+  pages = {16--35},
+  issn = {00405809},
+  doi = {10.1016/j.tpb.2025.08.004},
+  langid = {english}
+}
+
+@article{bezanson2017,
+  title = {Julia: {{A Fresh Approach}} to {{Numerical Computing}}},
+  shorttitle = {Julia},
+  author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+  date = {2017-01},
+  journaltitle = {SIAM Review},
+  shortjournal = {SIAM Rev.},
+  volume = {59},
+  number = {1},
+  pages = {65--98},
+  issn = {0036-1445, 1095-7200},
+  doi = {10.1137/141000671},
+  langid = {english}
+}
+
+@book{decker2025,
+  title = {The {{Computer Algebra System OSCAR}}: {{Algorithms}} and {{Examples}}},
+  shorttitle = {The {{Computer Algebra System OSCAR}}},
+  editor = {Decker, Wolfram and Eder, Christian and Fieker, Claus and Horn, Max and Joswig, Michael},
+  date = {2025},
+  series = {Algorithms and {{Computation}} in {{Mathematics}}},
+  publisher = {Springer Nature Switzerland},
+  location = {Cham},
+  issn = {1431-1550, 2512-3254},
+  doi = {10.1007/978-3-031-62127-7},
+  isbn = {978-3-031-62126-0 978-3-031-62127-7},
+  langid = {english}
+}
+
+@article{fenichel1979,
+  title = {Geometric Singular Perturbation Theory for Ordinary Differential Equations},
+  author = {Fenichel, Neil},
+  date = {1979-01},
+  journaltitle = {Journal of Differential Equations},
+  shortjournal = {Journal of Differential Equations},
+  volume = {31},
+  number = {1},
+  pages = {53--98},
+  issn = {00220396},
+  doi = {10.1016/0022-0396(79)90152-9},
+  langid = {english}
+}
+
+@thesis{goeke2013,
+  title = {Reduktion Und Asymptotische {{Reduktion}} von {{Reaktionsgleichungen}}},
+  author = {Goeke, Alexandra},
+  date = {2013},
+  institution = {RWTH Aachen},
+  location = {Aachen},
+  url = {https://publications.rwth-aachen.de/record/229008},
+  keywords = {Reaktionsgleichung (SWD) / Reduktion (SWD) / Asymptotik (SWD)}
+}
+
+@incollection{goeke2013a,
+  title = {Quasi-{{Steady State}}: {{Searching}} for and {{Utilizing Small Parameters}}},
+  shorttitle = {Quasi-{{Steady State}}},
[…]
