### 2026-05-07 29022a1
Remove paper files (JOSS review)
