+++ paper.md
@@ -1,225 +1,0 @@
----
-title: 'TikhonovFenichelReductions.jl: A systematic approach to geometric singular perturbation theory'
-tags:
-  - Julia
-  - Tikhonov
-  - Fenichel
-  - singular perturbation theory
-  - time scale separations 
-  - dynamical systems
-
-authors:
-  - name: Johannes Apelt
-    affiliation: 1
-affiliations:
- - name: Institute of Mathematics and Computer Science, University of Greifswald, Germany
-   index: 1
-date: 14 January 2026
-bibliography: paper.bib
----
-
-# Summary
-Natural systems often evolve on different characteristic time scales, i.e. the
-rates of processes differ substantially in orders of magnitude [@hek2010].
-Singular perturbation theory is a mathematical toolbox that translates this
-feature to ODE models of such systems and allows to perform dimensionality
-reduction when only one characteristic time scale (typically the slow one
-describing the long term behaviour) is of interest [@wechselberger2020].
-This is reflected by the presence of a small parameter $\varepsilon>0$, which
-can be interpreted as the conversion rate between the two time scales.
-Roughly speaking, the dynamics in slow time $\tau=\varepsilon t$ can be
-approximated by substituting the fast part of the system with its steady state,
-i.e.\ for the slow processes the fast changes appear instant.
-More precisely, we obtain the *reduced system* (or *reduction* in short) as the
-limit
-\begin{equation}
-  \label{eq:tikhonov}
-  \begin{aligned}
-    \dot{u}             & = g(u,v) \\
-    \varepsilon \dot{v} & = h(u,v)
-  \end{aligned}
-\quad\xrightarrow[\varepsilon\to0]{}\quad
- \begin{aligned}
-    \dot{u} & = g(u,v), \quad u\in\mathbb{R}^s \\ 
-    0       & = h(u,v), \quad v\in\mathbb{R}^r 
-  \end{aligned}
-\end{equation}
-with $\dot{}=\text{d}/\text{d}\tau$ as stated in Tikhonov's theorem
-[-@tikhonov1952; Theorem 1.1 in @verhulst2007],
-which is defined on the so-called slow manifold
-$M_0=\{(u,v)\,\vert\,h(u,v)=0\}$.
-For autonomous systems the results by @fenichel1979 allow a clear geometric
-interpretation --- hence the name geometric singular perturbation theory (GSPT)
---- and guarantee that for $\varepsilon>0$ sufficiently small the reduction
-captures the behaviour of the full system. 
-
-In practice GSPT can mitigate the realism-complexity trade-off, as modellers can
-work with the reduction instead of the full system, which is likely more
-feasible. 
-Because the reduction is still embedded in the full system as a special case
-defined by the time scales, one retains the original interpretability.
-This allows to follow the modelling paradigm: 
-Start with a complex but realistic model, reduce later.
-
-The algebraic approach recently developed by Goeke \& Walcher (and colleagues)
-allows to work with GSPT systematically and is coordinate-free, i.e. we do not
-rely on the standard form (\ref{eq:tikhonov}) with a time scale separation of
-components.
-Instead, we consider systems of the form 
-\begin{equation}
-    \label{eq:slowfastsystem}
-    x' = f^{(0)}(x) + \varepsilon f^{(1)}(x) + \mathcal{O}(\varepsilon^2), 
-    \quad x\in\mathbb{R}^n,
-\end{equation}
-where $f^{(0)}$ contains the fast processes and $f^{(1)}$ the slow ones.
[…]
