### 2026-05-26 d2c41a4
improved output if there exists no reduction
