### 2026-05-26 28b2473
fixed issue with wrong dimension of reduction if it is changed manually
