### 2026-05-26 c39532e
added progress bar and possibility to skip TFPV candidates
