+++ Project.toml
@@ -1,7 +1,10 @@
 name = "TikhonovFenichelReductions"
 uuid = "66629163-d46e-44ca-9a00-7025b1e168e6"
+version = "0.3.5"
 authors = ["jo-ap"]
-version = "0.3.4"
+
+[workspace]
+projects = ["docs", "test"]
 
 [deps]
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
@@ -10,6 +13,7 @@ Latexify = "23fbe1c1-3f47-55db-b15f-69d7ec21a316"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 Oscar = "f1435218-dba5-11e9-1e4d-f1a5fab5fc13"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 
 [weakdeps]
 Catalyst = "479239e8-5488-4da2-87a7-35f2df7eef83"
@@ -25,7 +29,5 @@ Latexify = "0.16"
 Logging = "1.9"
 Oscar = "1.4"
 Pkg = "1.9"
+ProgressMeter = "1.11.0"
 julia = "1.9"
-
-[workspace]
-projects = ["docs", "test"]
