### 2026-06-03 7e0be24
added orcid; minor changes
