+++ paper/paper.bib
@@ -45,7 +45,7 @@
   editor = {Decker, Wolfram and Eder, Christian and Fieker, Claus and Horn, Max and Joswig, Michael},
   date = {2025},
   series = {Algorithms and Computation in Mathematics},
-  publisher = {Springer Nature Switzerland},
+  publisher = {Springer},
   location = {Cham},
   issn = {1431-1550, 2512-3254},
   doi = {10.1007/978-3-031-62127-7},
@@ -213,9 +213,11 @@
 }
 
 @software{oscar2022,
+  key = {OSCAR},
+  organization = {The OSCAR Team},
   title = {{OSCAR} -- {Open Source Computer Algebra Research} System},
-  author = {The OSCAR Team},
-  date = {2025},
+  shorttitle = {{OSCAR}},
+  date = {2026},
   url = {https://www.oscar-system.org}
 }
 
